requirements.txt
pyTelegramBotAPI==4.10.0
pydantic==1.10.2
openai==0.27.2
pysqlite3==0.5.0
pydub==0.25.1
ffmpeg
google_cloud_texttospeech==2.14.1
google-auth 
google-auth-oauthlib